requirements.txt
flask==2.3.3
werkzeug==2.3.7
psutil==5.9.5
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pydantic==2.4.2
typing-extensions==4.8.0
PyYAML==6.0.1
